feat: Update email sender in mark_application_as_passed view
The `mark_application_as_passed` view in `views.py` has been updated to use the `DEFAULT_FROM_EMAIL` setting instead of `EMAIL_HOST_USER` for the sender email address in the `send_mail` function. This change ensures that the email is sent from the default email address specified in the project settings.