refactor: Show spinner when adding a skill in job_detail_include.html
This commit adds a spinner to the "Add Skill" button in the job_detail_include.html template. When the button is clicked, the spinner is displayed to indicate that the skill is being added. This improves the user experience by providing visual feedback and preventing multiple clicks while the skill is being processed.