refactor: Fix slug generation in AddJobLink view
This commit fixes the slug generation in the AddJobLink view in views.py. Previously, the slug was not being properly generated using the role title and company name. The code has been updated to correctly generate the slug using the slugify function. This change ensures that the slug is accurate and follows the desired format, improving the functionality of the application.