refactor: Update company_detail.html to use job role as default title
This commit updates the company_detail.html template to use the job role as the default title if it is available. Previously, only the job title was used as the title for the page. This change ensures that the job role is displayed in the title if it exists, providing more descriptive information to users.